feat(library): Rekordbox ANLZ ingest + auto-cue engine + section builder
Lands the v8.2 library expansion. The library subpackage grows a
Rekordbox-grounded ingest path, an offline cue-detection engine, a
section/transition builder shared with intel/, and the supporting
smart-cue policy + model-asset surface.

New modules (src/vibemix/library/):
  anlz_ingest          — Rekordbox ANLZ phrase/cue/beatgrid reader
                         (no rekordbox-running requirement, no SQLCipher);
                         the path picked after spike-003 verdict that
                         PRO DJ LINK is dead laptop-only and master.db
                         is history-only.
  audio_decode         — PyAV/FFmpeg-only decode helper (torch-free)
  audio_features       — narrow numpy DSP helpers used by cue_engine
  cache_paths          — single source of truth for ~/.cache/vibemix/*
                         paths (per-backend namespacing)
  cue_detr             — CUE-DETR onnx inference seam (server-side
                         path lives on bravoh-gpu-worker; fallback
                         to heuristic when absent)
  cue_engine           — orchestrator: ANLZ floor + heuristic fallback
                         + (optional) CUE-DETR refine
  cue_refine           — heuristic refiner over the producer's emit
  cue_types            — CueAnchor shared dataclass (seam to intel/)
  embed_cache          — content-hash cache rows keyed by backend
  embed_config         — model-asset path resolution
  embed_factory        — builds the embedder backend per config
                         (default onnx CLAP, torch reference kept
                         only for parity validation)
  embed_types          — Embedder + ClapEmbedder protocols
  model_assets         — discovery + readiness for required local
                         model assets (CLAP onnx, sqlite-vec, etc.)
  section_builder      — derives Section + bars_until_section_end
                         from ANLZ phrase data, shared with
                         intel/transition_scorer.
  smart_cues           — A–H cue-slot policy + Rekordbox export marks

Test additions:
  tests/library/test_anlz_ingest, test_audio_decode, test_cache_paths,
  test_cue_engine, test_cue_refine, test_ingest_cli_anlz, test_smart_cues,
  test_models_cli — 7 new test files; plus updates to next_suggestion,
  toolset, setprep_tools, search, similar, sequencer, embed_*, telegram,
  store, ingest, etc. to wire the new modules in.

Removed:
  library/agent.py + tests/library/test_agent.py — superseded by the
  Codex+MCP curator path (toolset.py + codex_curate.py + mcp_server.py)
  library/youtube_ingest.py + tests/library/test_youtube_ingest.py —
  reverted; deferred to Bravoh per v8.2 milestone closeout

Verification: 589 library tests pass, 1 xfail (the known €50 budget
gate, see test_budget.py docstring).

+++ src/vibemix/library/export_rekordbox.py
@@ -200,7 +200,7 @@ def _add_track(xml: Any, item: Mapping[str, Any], filepath: Any) -> Any:
     duration_s = item.get("duration_s")
     if duration_s is not None:
         try:
-            kwargs["TotalTime"] = int(round(float(duration_s)))
+            kwargs["TotalTime"] = round(float(duration_s))
         except (TypeError, ValueError):
             pass
